merge

+++ client/src/components/GameSettings.jsx
@@ -2,6 +2,7 @@ import { useState, useRef, useEffect } from "react";
 import { MdMusicNote, MdMusicOff } from "react-icons/md";
 import { Link, useLocation, useNavigate } from "react-router-dom";
 import backgroundMusic from "../assets/audio/Shortwire - Reconfig [PV].mp3";
+import Swal from "sweetalert2";
 
 function GameSettings() {
   const [music, setMusic] = useState(true);
@@ -11,16 +12,29 @@ function GameSettings() {
   const navigate = useNavigate();
 
   const signOut = () => {
-    localStorage.clear();
-    navigate("/");
+    Swal.fire({
+      title: "<h2>Are you sure?</h2>",
+      icon: "warning",
+      html: "You will be logged out",
+      showCancelButton: true,
+      focusConfirm: false,
+      confirmButtonText: "<span>Yes</span>",
+      confirmButtonColor: "#F26379",
+      cancelButtonText: "<span>Cancel</span>",
+      cancelButtonColor: "#6D27D9",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        navigate("/");
+      }
+    });
   };
 
   const toggleMusic = () => {
     setMusic(!music);
   };
 
-  useEffect(() => {
-  }, [location.pathname]);
+  useEffect(() => {}, [location.pathname]);
 
   useEffect(() => {
     // Memainkan atau memberhentikan musik latar belakang saat komponen dimuat atau di-unmount
@@ -56,7 +70,7 @@ function GameSettings() {
             </div>
           )}
           <div>
-            {location.pathname !== "/" && (
+            {location.pathname !== "/" && location.pathname !== '/main-menu' && (
               <button onClick={signOut}>Quit Game </button>
             )}
           </div>
